academic/covid-xray-detection/requirements.txt
# Core ML/DL
tensorflow>=2.15.0
numpy>=1.24.0
scikit-learn>=1.2.0

# Data processing
pandas>=2.0.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tabulate>=0.9.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Jupyter (optional)
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: Web demo
# gradio>=3.50.0
# streamlit>=1.28.0
